LayoutTests/fast/element-targeting/prefer-succinct-selectors-expected.txt

PASS firstTarget is "#overlay"
PASS secondTarget is "DIV.fixed"
PASS successfullyParsed is true

TEST COMPLETE
This test requires WebKitTestRunner.
